fix: moved Info and Result to utils, warmstarting
