+++ hydra_plugins/hypersweeper/__init__.py
@@ -1,5 +1,6 @@
 from .hypersweeper import Hypersweeper
 from .hypersweeper_backend import HypersweeperBackend
-from .hypersweeper_sweeper import HypersweeperSweeper, Info, Result
+from .hypersweeper_sweeper import HypersweeperSweeper
+from .utils import Info, Result
 
 __all__ = ["HypersweeperSweeper", "HypersweeperBackend", "Info", "Result", "Hypersweeper"]
